+++ src/types/auth-signup/i-GenderState.ts
@@ -1,0 +1,3 @@
+export interface IGenderState {
+  gender: string;
+}
